requirements.txt
aiofiles==0.8.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
chromedriver-binary==125.0.6403.0.0
cinemagoer==2023.5.1
click==8.1.7
dataclasses-json==0.6.4
deeppyer==1.0.1
discord.py>=2.3.2
dnspython==2.6.1
frozenlist==1.4.1
giphpy==0.1.2
googlesearch-python==1.2.3
greenlet==3.0.3
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.6
IMDbPY==2022.7.9
isodate==0.6.1
joblib==1.3.2
lxml==5.2.1
malclient-upgraded==1.3.3
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
osrparse==6.0.2
ossapi==3.4.4
outcome==1.3.0.post0
packaging==24.0
pafy==0.5.5
pandas==2.2.1
pillow==10.3.0
prawcore==2.4.0
pydantic==1.10.15
pyimgbox==1.0.7
pymongo==4.6.3
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-youtube==0.9.4
pytz==2024.1
qrcode==7.4.2
randfacts==0.21.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rule34==1.8.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
textblob==0.18.0.post0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.11.0
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
vaderSentiment==3.3.2
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
youtube-dl==2021.12.17
